+++ app/controllers/videos_controller.rb
@@ -4,8 +4,7 @@
 class VideosController < ApplicationController
   before_action :authenticate_user!
   before_action :check_system_admin
-  before_action :set_video,               only: [:show, :edit, :update, :destroy]
-  before_action :redirect_without_video,  only: [:show, :edit, :update, :destroy]
+  before_action :find_video_or_redirect, only: [:show, :edit, :update, :destroy]
 
   # GET /videos
   def index
@@ -20,7 +19,7 @@ class VideosController < ApplicationController
   # POST /videos/save_video_order.js
   def save_video_order
     params[:video_ids].each_with_index do |video_id, index|
-      video = Video.current.where(video_page: params[:video_page]).find_by_id(video_id)
+      video = Video.current.where(video_page: params[:video_page]).find_by(id: video_id)
       video.update position: index if video
     end
     head :ok
@@ -43,7 +42,7 @@ class VideosController < ApplicationController
   def create
     @video = current_user.videos.new(video_params)
     if @video.save
-      redirect_to @video, notice: 'Video was successfully created.'
+      redirect_to @video, notice: "Video was successfully created."
     else
       render :new
     end
@@ -52,7 +51,7 @@ class VideosController < ApplicationController
   # PATCH /videos/1
   def update
     if @video.update(video_params)
-      redirect_to @video, notice: 'Video was successfully updated.'
+      redirect_to @video, notice: "Video was successfully updated."
     else
       render :edit
     end
@@ -61,13 +60,14 @@ class VideosController < ApplicationController
   # DELETE /videos/1
   def destroy
     @video.destroy
-    redirect_to videos_path, notice: 'Video was successfully deleted.'
+    redirect_to videos_path, notice: "Video was successfully deleted."
   end
 
   private
 
-  def set_video
-    @video = Video.current.find_by_id(params[:id])
+  def find_video_or_redirect
+    @video = Video.current.find_by(id: params[:id])
+    redirect_without_video
   end
 
   def redirect_without_video
@@ -75,6 +75,9 @@ class VideosController < ApplicationController
   end
 
   def video_params
-    params.require(:video).permit(:video_page, :vimeo_number, :photo, :photo_cache, :position, :archived)
+    params.require(:video).permit(
+      :video_page, :vimeo_number, :client_name, :agency_name,
+      :photo, :photo_cache, :position, :archived
+    )
   end
 end