+++ db/migrate/20170520195848_add_client_name_and_agency_name_to_videos.rb
@@ -1,0 +1,6 @@
+class AddClientNameAndAgencyNameToVideos < ActiveRecord::Migration[5.1]
+  def change
+    add_column :videos, :client_name, :string
+    add_column :videos, :agency_name, :string
+  end
+end